Add parse_promise_summary function
Remove the logic of out of the main function which parses the contents of the
promise summary into the parse_promise_summary function. Update the function to
returns a dict containing the run start and end time, the version string, the
percentage of promises kept and repaired, the last line of the log file and the
path to the log itself.

Move the summary logic which prints the results to the screen into a function
called display_summary. This keeps the default behavior of this script to
unchanged.